Fixing Styling

+++ 0x00-python_variable_annotations/7-to_kv.py
@@ -1,8 +1,8 @@
 #!/usr/bin/env python3
-from typing import Union
 """type annotation"""
+from typing import Union
 
 
-def to_kv(k: str, v: Union[int, float]) -> tuple:
+def to_kv(k: str, v: Union[float, int]) -> tuple:
     """returns a tuple"""
     return (k, v ** 2)
